feat: comprehensive upgrade of core agent architecture, sidecar parsers, and win automation
This commit includes:
- Upgraded phantom-core agent registry, prompt firewall, identity, hotkey handlers, and intent gates.
- Refactored kairo-sidecar parsers (docx, xlsx, pdf) and writers (docx, xlsx, pptx).
- Added multi-layer integration tests, gauntlet reports, and validation datasets.
- Enhanced Windows scenario automations (Word, Excel, PPTX, VSCode, Obsidian, Notepad).
- Configured local model finetuning and dspy prompt optimizer setups.

+++ phantom-core/src/screen_context.rs
@@ -155,7 +155,7 @@ impl ScreenContextEngine {
                     biHeight: -height, // Top-down
                     biPlanes: 1,
                     biBitCount: 24,
-                    biCompression: BI_RGB.0 as u32,
+                    biCompression: BI_RGB.0,
                     ..Default::default()
                 },
                 ..Default::default()
